requirements.txt
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
dnspython==1.16.0
eventlet==0.30.2
Fiona==1.8.19
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
geopandas==0.9.0
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.20.2
pandas==1.2.4
pyproj==3.0.1
python-dateutil==2.8.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rq==1.8.0
rtree==0.9.7
Shapely==1.7.1
six==1.15.0
urllib3==1.26.5
websocket-client==0.58.0
Werkzeug==1.0.1
xlrd==2.0.1
distinctipy==1.1.5
matplotlib-scalebar==0.7.2
openpyxl==3.0.7
-e .
